GearSwap v0.821 - Initial Commit
Renamed out_arr to command_registry and basically rewrote all of the
logic handling index creation/deletion. It is now keyed to the timestamp
of the time when GearSwap was first able to calculate a complete packet
for the action. This should avoid having midaction() reset itself when
someone spams the same spell.