Battlemod: Fix missing "KO" status
Anything that caused a KO status (i.e. Odin) was giving an "ERROR 120" status message instead.
This was because Battlemod was ignoring status=0, which is the value of KO.

I've removed the erroneous check, which should put KO back on the list of statuses that can be given via actions/add effects/spikes.