Create irc.lua
This is an irc chat based addon. It works with twitch.tv and should work with most other irc servers. Currently it is not possible to set the chat to default to the irc server but a functionality like /cm i will be added in the future to make it act like a normal Linkshell.

Known issues: The games autotranslate text causes problems when sent to the IRC server.