organizer now prefers putting gear in wardrobe over inventory
If an item can be stored in the wardrobe, that is, it is equippable (it
does not matter whether the item was mentioned in the sets or the
organizer_items variable), it is now added to the wardrobe. All
non-equippable items are added to the inventory directly. If items
stored in the wardrobe pass the wardrobe max limit, the surplus items
are moved into the inventory.  If the inventory max limit is passed
after surplus items from wardrobe has been moved into inventory, an
error message is given. The reason for all of this logic is to use the
wardrobe in a more effective manner, while leaving more space for loot
in the inventory.