texts library - Exposing m.name
It was impossible to associate the windower.text official name of a text
with the texts library table reference of the text, which makes it
impossible to delete texts in a way that is recognized by the library if
you don't know the table handle. This is one solution to that problem.