Update Pouches.lua
Updated pouches to check if your inventory is full before 

Having pouches try and fail to use items when your inventory is full is mildly problematic. No real player is going to sit there for minutes to hours opening pouches they can't use. If SE decided to get strict this is a good check.