Update gearswap.lua
The functionality to resolve target IDs passed to commands in the chatlog, upon which the send command relies for the <tid> parameters to work, only works when gearswap has an user profile loaded and active. This leads to many complaints that the functionality is not working, simply because it's being used on a character without gearswap, or without a profile loaded.

This change will make gearswap create an empty default.lua on load, which is sufficient for it to become active and start resolving target IDs.

(I am not sure about the wording of the notice, and of the contents of the empty file. I am also not sure about the place - doing it on load feels the most appropriate. I'm open to suggestions on all of these.)