Fixes player.mp error on load
When gearswap is loaded, it is initializing stuff basically by re-parsing stuff from windower.packets.last_incoming() by looping over its handlers for various packets.
This includes the last 0x028 which relies on knowing the players MP (for checking if you have enough mp to cast the spell). So if it parses the last 0x028 before it parses 0x0DF or 0x0E2 then an error occurs. So this solution just ignores the last 0x028.