Update jobchange.lua
This is a change to job change to allow it to work consistently in Odyssey. 

The first time you enter Idyssey the server will accept an `0x100`, but once you initiate an NM fight the sever will no longer accept an `0x100` until you poke the Moogle again without the BCNM flag. I added a poke for Nomad Moogles as well since technically there is no manual way to change jobs without poking a Nomad Moogle, and if SE wanted to get banhappy it's something they could easily look for.

Note -- changing jobs manually the first time you enter the BC, the last two bytes(unknown 1 & 2) are empty, but once you've killed an NM -- they do contain *some* values, but they don't seem to be related to anything related to the event. The server still accepts a traditional `0x100`.