Items with - and + are not being pulled out of bags
Looks like any item with a hyphen (-) or plus sign (+) is not being pulled out of storage by this lua when these items are called via keybinds, macros, or commands.

The imatch function uses gsub and in lua 5.1 this requires these special characters to be preceded by an escape character, %.  I've added two lines to replace these characters with the appropriate text to find them on lookup.

I'm not the most familiar with the string and table libraries of windower so there may be a better way to do this but I'm hoping this at least highlights the issue and provides a solution so someone may improve upon it.