Fixes bug around /bstpet prefix.
In-game, /bstpet accepts both a number (an index of the pet move to perform) or the full name of the desired move. Previously gearswap would only trigger when the index was provided and now with the full name. This change fixes triggers.lua to trigger gearswap when /bstpet is used with the pet ability name.