[GearSwap] Remove unnecessary command registry spell assignment
For each action which triggers GearSwap, there is only ever one spell table in
memory for that action (from trigger -> midcast). cmd_reg:new_entry already
assigns a reference to that spell table when the command registry entry is
created, so there is no need to assign the same reference again.

Small note: a new, duplicate spell table is created post-midcast in response to
receiving an incoming action packet, so if that action packet was received as a
result of a GearSwap fired action then technically there are now two spell
tables in memory for the same action. But:
1. We don't care that much because the critical task of firing the action
correctly was already completed successfully.
2. If we really wanted to we could reuse the original spell table and avoid
creating a second one, only allocating memory for a new spell table if the
original spell table could not be retrieved.