[GearSwap] Prevent arithmetic errors with unvalidated target IDs
In outgoing text, if a numerical target ID is provided as part of the command,
and the command contains a legal action, it will pass through the
filter_pretarget and tonumber guard clauses and end up in the remaining code
path. At this point, we can see we are running head first into packet assembly
without having done any validation on spell.target.id and spell.target.index,
both of which are used in packet assembly and both of which could be nil if the
get_mob_by_id call in valid_target returned nil. In such an event, an
arithmetic error will occur during packet assembly, '/ma "Dia" 12345' will
quickly reproduce this (if you know the Dia spell).

It should also be noted, that in this scenario since we did not take the
filtered_action code path, nor the pretarget code path, there is only one code
path remaining that we can be on, which is the precast code path.

Enter filter_precast, the gatekeeper of the precast code path. filter_precast
exists only for a single purpose, to perform exactly the validation we are
missing in this scenario before we can proceed with packet assembly.

While we are in the vicinity, as a quick aside, commands such as
'/ma "Dia" <t> 123 456' are valid and will be accepted by both the vanilla
client and GearSwap, and will result in the spell being cast. For this reason
we cannot be assured the last item in splitline is the action's target, so
spell.target.raw will be used instead.