[GearSwap] Move st pronoun check to appropriate place and fix small bug
We should check for an st pronoun at the earliest possible moment, i.e.
immediately after a potential target change, since there can never be an action
that gets fired for an st pronoun. By definition, an st pronoun will trigger a
subtarget arrow for the user to select a target with, so there is no target
until that selection is made and a new command is issued to outgoing text with
the selected target. For this reason it makes sense to check for an st pronoun
before doing any packet assembly, because those packets will never be used in
case of an st pronoun. Furthermore, there is nothing else to do after setting
st_flag except wait for the 'real' command with the real target to be received
by outgoing text, so we should return immediately.

The previous st pronoun code path would return nil after setting st_flag, but
this is not correct, since the original target could have been <t> but the user
pretarget function decided to change this to <st>. In this case GearSwap must
return the <st> pronoun to outgoing text in order to grant the user the
subtarget arrow they requested. However at present since nil is returned, it is
<t> that will get released to outgoing text instead of <st>, which will cause
several problems:
1. The action will be initiated immediately with the current target, which could
   be both the unintended target and the unintended timing.
2. There will be no precast or midcast gear swaps for the action.
3. The st_flag is left set, which means the next valid command GearSwap receives
   will also not receive any gear swaps since outgoing text will return early
   after clearing the st_flag before initiating standard GearSwap flow.

Since this commit is a more innocent change, did all code style updates for
equip_sets_exit in this commit.