[GearSwap] Only do item trade check once per code path
At present for the pretarget code path, the bit.band spawn type check gets
performed twice, once in outgoing text, and again in pretarget. We should only
need to perform this check once. In addition, the determination if the current
action type is a trade or not is only performed in outgoing text, before the
userfile might change the target to something different in pretarget. If the
target did get changed in pretarget to something with a different spawn type,
the initial trade action type that was determined in outgoing text might no
longer be correct. This would be the case if the initial target was an NPC, and
the new target was non-NPC, or vice versa. So this commit determines the item
action type for the pretarget code path only after the target may have been
changed to ensure it is correct, and moves the outgoing text item action type
check to inside the target ID code path to avoid it being performed twice for
pretarget code path.